Reorganize mentions code
