+++ com/notifications.js
@@ -5,7 +5,8 @@ const renderAvatar = require('./avatar')
 const renderName = require('./name')
 const renderPostActions = require('./feed/post-actions')
 const renderHeartIcon = require('./icons/heart')
-const {timestamp, linkifyText} = require('../lib/util')
+const {timestamp} = require('../lib/util')
+const {linkifyText} = require('../lib/posts')
 
 // exported api
 // =
