+++ index.less
@@ -9,6 +9,7 @@
 @import "./com/feed/post-actions";
 @import "./com/feed/thread";
 @import "./com/feed/show-posts-btn";
+@import "./com/feed/mention-window";
 @import "./com/profile/friends-list";
 @import "./com/profile/profile-card";
 @import "./com/profile/profile-header";
@@ -16,7 +17,6 @@
 @import "./com/profile/following";
 @import "./com/crop-popup";
 @import "./com/who-to-follow";
-@import "./com/mention-window";
 
 @import "./views/feed";
 @import "./views/settings";
